feat(server): `pnpm server` becomes `pnpm server start`
SEMI-BREAKING CHANGE:

`pnpm server` doesn't start a server but only prints help

`pnpm server start` starts a store server
